[REF] web_tour: move MacroMutationObserver
In this commit, we move class MacroMutationObserver into
tour_interactive.
The MacroMutationObserver class no longer has anything to do with the
Macro class and is now used ONLY in tour_interactive.
The MacroMutationObserver is renamed with TourInteractiveObserver
