+++ addons/web_tour/static/src/js/tour_interactive/tour_interactive_observer.js
@@ -1,0 +1,76 @@
+export class TourInteractiveObserver {
+    observerOptions = {
+        attributes: true,
+        childList: true,
+        subtree: true,
+        characterData: true,
+    };
+    constructor(callback) {
+        this.callback = callback;
+        this.observer = new MutationObserver((mutationList, observer) => {
+            callback(mutationList);
+            mutationList.forEach((mutationRecord) =>
+                Array.from(mutationRecord.addedNodes).forEach((node) => {
+                    let iframes = [];
+                    if (String(node.tagName).toLowerCase() === "iframe") {
+                        iframes = [node];
+                    } else if (node instanceof HTMLElement) {
+                        iframes = Array.from(node.querySelectorAll("iframe"));
+                    }
+                    iframes.forEach((iframeEl) =>
+                        this.observeIframe(iframeEl, observer, () => callback())
+                    );
+                    this.findAllShadowRoots(node).forEach((shadowRoot) =>
+                        observer.observe(shadowRoot, this.observerOptions)
+                    );
+                })
+            );
+        });
+    }
+    disconnect() {
+        this.observer.disconnect();
+    }
+    findAllShadowRoots(node, shadowRoots = []) {
+        if (node.shadowRoot) {
+            shadowRoots.push(node.shadowRoot);
+            this.findAllShadowRoots(node.shadowRoot, shadowRoots);
+        }
+        node.childNodes.forEach((child) => {
+            this.findAllShadowRoots(child, shadowRoots);
+        });
+        return shadowRoots;
+    }
+    observe(target) {
+        this.observer.observe(target, this.observerOptions);
+        //When iframes already exist at "this.target" initialization
+        target
+            .querySelectorAll("iframe")
+            .forEach((el) => this.observeIframe(el, this.observer, () => this.callback()));
+        //When shadowDom already exist at "this.target" initialization
+        this.findAllShadowRoots(target).forEach((shadowRoot) => {
+            this.observer.observe(shadowRoot, this.observerOptions);
+        });
+    }
+    observeIframe(iframeEl, observer, callback) {
+        const observerOptions = {
+            attributes: true,
+            childList: true,
+            subtree: true,
+            characterData: true,
+        };
+        const observeIframeContent = () => {
+            if (iframeEl.contentDocument) {
+                iframeEl.contentDocument.addEventListener("load", (event) => {
+                    callback();
+                    observer.observe(event.target, observerOptions);
+                });
+                if (!iframeEl.src || iframeEl.contentDocument.readyState === "complete") {
+                    callback();
+                    observer.observe(iframeEl.contentDocument, observerOptions);
+                }
+            }
+        };
+        observeIframeContent();
+        iframeEl.addEventListener("load", observeIframeContent);
+    }
+}
